+++ front-end/Android-app/Grafia/app/src/main/java/com/example/spider/grafia/LoginActivity.kt
@@ -1,11 +1,9 @@
 package com.example.spider.grafia
 
-import android.content.Context
 import android.content.Intent
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
 import kotlinx.android.synthetic.main.activity_login.*
-import java.io.File
 import java.lang.Exception
 import java.lang.StringBuilder
 import java.security.MessageDigest
