+++ front-end/Android-app/Grafia/app/src/main/java/com/example/spider/grafia/VideoActivity.kt
@@ -109,7 +109,6 @@ class VideoActivity : AppCompatActivity() {
         }
 
         uploadVideo.setOnClickListener {
-            UploadFileAsync(projectPath).execute("")
         }
 
 
@@ -158,33 +157,7 @@ class VideoActivity : AppCompatActivity() {
                 myFile.delete()
                 mCurrentVideoPath = ""
             }
-            //cambiar esto
             val videoURI = data?.data
-
-            val wholeID = DocumentsContract.getDocumentId(videoURI)
-
-            // Split at colon, use second item in the array
-            val id = wholeID.split(":".toRegex()).dropLastWhile { it.isEmpty() }.toTypedArray()[1]
-
-            val column = arrayOf(MediaStore.Video.Media.DATA)
-
-            // where id is equal to
-            val sel = MediaStore.Video.Media._ID + "=?"
-
-            val cursor = this.getContentResolver().query(MediaStore.Video.Media.EXTERNAL_CONTENT_URI, column, sel, arrayOf(id), null)
-
-            var filePath = ""
-            val columnIndex = cursor.getColumnIndex(column[0])
-
-            if (cursor.moveToFirst()) {
-                filePath = cursor.getString(columnIndex)
-            }
-            cursor.close()
-
-            mCurrentPickedVideo = filePath
-
-            val timeStamp: String = java.text.SimpleDateFormat("yyyyMMdd_HHmmss").format(java.util.Date())
-            mCurrentPickedVideoName = "VIDEO_${userId}_${timeStamp}_.jpg"
 
             if (videoURI != null) {
                 mCurrentVideoUri = videoURI
@@ -283,54 +256,4 @@ class VideoActivity : AppCompatActivity() {
         }
         return file.delete()
     }
-
-
-
-    private inner class UploadFileAsync(val projectPath: String) : AsyncTask<String, Void, String>() {
-
-        override fun doInBackground(vararg params: String): String {
-
-            var path = ""
-            var name = ""
-            if (mCurrentVideoPath == "" && mCurrentPickedVideo != "") {
-                path = mCurrentPickedVideo
-                name = mCurrentPickedVideoName
-
-            } else if (mCurrentVideoPath != "" && mCurrentPickedVideoName == "" ) {
-                path = mCurrentVideoPath
-                name = path.substringAfterLast("/")
-            }
-
-
-            val multipart = Multipart(URL("http://54.81.239.120/fUploadAPI.php"))
-            multipart.addFormField("fileType", "1")
-            multipart.addFormField("path", (projectPath + "/videos/"))
-            multipart.addFilePart("file", path, name, "video/mp4")
-
-            val bool = multipart.upload()
-
-            if (bool) {
-                return "YES"
-            } else {
-                return "NO"
-            }
-
-        }
-
-        override fun onPostExecute(result: String) {
-
-            if (result == "YES") {
-                Toast.makeText(this@VideoActivity, "Uploaded!", Toast.LENGTH_LONG).show()
-            } else {
-                Toast.makeText(this@VideoActivity, "Try Again", Toast.LENGTH_LONG).show()
-            }
-
-        }
-
-        override fun onPreExecute() {}
-
-        override fun onProgressUpdate(vararg values: Void) {}
-    }
-
-
 }
