+++ front-end/Android-app/Grafia/app/src/main/res/layout/activity_login.xml
@@ -34,7 +34,7 @@
             android:layout_marginStart="31dp"
             app:layout_constraintEnd_toEndOf="parent"
             android:layout_marginEnd="286dp" app:layout_constraintHorizontal_bias="0.482"
-            android:layout_marginTop="31dp" app:layout_constraintTop_toBottomOf="@+id/editText"/>
+            android:layout_marginTop="31dp" app:layout_constraintTop_toBottomOf="@+id/email"/>
     <TextView
             android:text="Email Address:"
             android:layout_width="wrap_content"
@@ -60,7 +60,7 @@
             android:layout_height="wrap_content"
             android:inputType="textEmailAddress"
             android:ems="10"
-            android:id="@+id/editText" app:layout_constraintEnd_toEndOf="parent"
+            android:id="@+id/email" app:layout_constraintEnd_toEndOf="parent"
             android:layout_marginEnd="18dp" android:layout_marginTop="19dp"
             app:layout_constraintTop_toBottomOf="@+id/textView" app:layout_constraintStart_toStartOf="parent"
             android:layout_marginStart="16dp" android:hint="Email"/>
